Add Pokemon raid module
This module is for managing a queue for Pokemon
Sword & Shield raids. The queue is a simple FILO
queue, so it should be relatively fair.
The queue is also auto rotating, so no need to
rejoin.

New commands are:
!raid - creates a new raid
!rj - lets a user join the raid queue
!raiders - lists the current queue
!r1, !r2, !r3 - to fetch 1, 2 or 3 users from the
queue. The users will be removed from the front of
the queue and reappended at the end.